+++ t_symlink_secure.c
@@ -1,8 +1,8 @@
 /*
- * Test harness for the fake-super branches of vfs_symlink_at()/do_mknod_at().
+ * Test harness for the fake-super branches of vfs_symlink_at()/vfs_mknod_at().
  * Fake-super stores a symlink/device as a placeholder file, so the create
  * resolves the final component; the no-slash branch used to fall back to
- * vfs_symlink()/do_mknod(), whose plain open() followed a planted basename
+ * vfs_symlink()/vfs_mknod(), whose plain open() followed a planted basename
  * symlink and escaped the module. Checks the fixed wrappers refuse it;
  * --poc shows the old fallback escaping. Not linked into rsync. GPL version 2.
  */
@@ -118,15 +118,15 @@ int main(int argc, char **argv)
 	am_root = -1;	/* fake-super: symlinks/devices stored as files */
 
 	if (poc) {
-		/* Pre-fix fallback: a no-slash path went to vfs_symlink()/do_mknod(),
+		/* Pre-fix fallback: a no-slash path went to vfs_symlink()/vfs_mknod(),
 		 * which open() the basename without O_NOFOLLOW. */
 #ifdef TEST_SYMLINK_PLACEHOLDER
 		vfs_symlink("VULN_SYM_PAYLOAD", "sympath");
 		check_clobbered("poc vfs_symlink bare", "../outside/secret_sym",
 				"VULN_SYM_PAYLOAD");
 #endif
-		do_mknod("nodpath", S_IFCHR | 0600, 0);
-		check_clobbered("poc do_mknod bare", "../outside/secret_nod", "");
+		vfs_mknod("nodpath", S_IFCHR | 0600, 0);
+		check_clobbered("poc vfs_mknod bare", "../outside/secret_nod", "");
 		return errs ? 1 : 0;
 	}
 
@@ -141,17 +141,11 @@ int main(int argc, char **argv)
 	check_preserved("vfs_symlink_at slashed", "../outside/secret_sym2", "VICTIM_SYM2");
 #endif
 
-# ifdef HAVE_MKNODAT
-	/* Without mknodat() do_mknod_at() IS do_mknod(): the confinement is
-	 * compiled out by design (SECURITY.md), so these would assert a
-	 * property the build deliberately does not have.  The do_symlink_at()
-	 * checks above do not depend on it and still run. */
-	do_mknod_at("nodpath", S_IFCHR | 0600, 0);
-	check_preserved("do_mknod_at bare", "../outside/secret_nod", "VICTIM_NOD");
+	vfs_mknod_at("nodpath", S_IFCHR | 0600, 0);
+	check_preserved("vfs_mknod_at bare", "../outside/secret_nod", "VICTIM_NOD");
 
-	do_mknod_at("sub/nodpath2", S_IFCHR | 0600, 0);
-	check_preserved("do_mknod_at slashed", "../outside/secret_nod2", "VICTIM_NOD2");
-# endif
+	vfs_mknod_at("sub/nodpath2", S_IFCHR | 0600, 0);
+	check_preserved("vfs_mknod_at slashed", "../outside/secret_nod2", "VICTIM_NOD2");
 
 	if (errs)
 		fprintf(stderr, "%d failure(s)\n", errs);
